feat: configurable media scanner host / port
